+++ webservice/GateWashDataService/Models/GroupedIncreaseModel.cs
@@ -12,13 +12,13 @@ namespace GateWashDataService.Models
         public string Terminal { get; set; }
         public string ProgramsDescription { get; set; }
         public string TypesDescription { get; set; }
-        public int Amount { get; set; }
+        public double Amount { get; set; }
         public int ProgramCount { get; set; }
         public decimal ARPU
         {
             get
             {
-                return this.Amount / this.ProgramCount;
+                return (decimal)this.Amount / this.ProgramCount;
             }
             private set { }
         }
